+++ lib/govuk_personalisation.rb
@@ -6,7 +6,4 @@ require "govuk_personalisation/flash"
 require "govuk_personalisation/test_helpers/features"
 require "govuk_personalisation/test_helpers/requests"
 
-module GovukPersonalisation
-  class Error < StandardError; end
-  # Your code goes here...
-end
+module GovukPersonalisation; end